requirements.txt
pandas
numpy
pytorch
holidays
gdown